Add the ability for plugins to opt-in to asyncronous state calls
This adds a new function on state-context that allows the plugin to
opt in to receiving save and load calls on a new "state-thread"
symbolic thread. Hosts could use this to parallelize loading and
saving session state, which can greatly reduce the time it takes to
load or save a large session.